+++ internal/provider/resource_http_request_test.go
@@ -7,10 +7,12 @@ import (
 	"context"
 	"encoding/base64"
 	"encoding/json"
+	"fmt"
 	"github.com/hashicorp/terraform-plugin-framework/diag"
 	fresource "github.com/hashicorp/terraform-plugin-framework/resource"
 	"github.com/hashicorp/terraform-plugin-framework/tfsdk"
 	"github.com/hashicorp/terraform-plugin-go/tftypes"
+	"github.com/hashicorp/terraform-plugin-testing/terraform"
 	"github.com/rios0rios0/terraform-provider-http/test/infrastructure/builders"
 	"github.com/stretchr/testify/assert"
 	"testing"
@@ -40,45 +42,65 @@ func TestHTTPRequestResource(t *testing.T) {
 	t.Parallel()
 
 	t.Run("should apply and check the state when using GET method", func(t *testing.T) {
-		var state1 bytes.Buffer
-		_ = json.Compact(&state1, []byte(`{
+		// given
+		var state bytes.Buffer
+		_ = json.Compact(&state, []byte(`{
 		"method": "GET",
 		"path": "/posts/1",
 		"response_code": 200,
 		"response_body": "{\n  \"userId\": 1,\n  \"id\": 1,\n  \"title\": \"sunt aut facere repellat provident occaecati excepturi optio reprehenderit\",\n  \"body\": \"quia et suscipit\\nsuscipit recusandae consequuntur expedita et cum\\nreprehenderit molestiae ut ut quas totam\\nnostrum rerum est autem sunt rem eveniet architecto\"\n}"
 	}`))
-		state1ID := base64.StdEncoding.EncodeToString(state1.Bytes())
+		stateID := "anything unique"
+		modelEncoded := base64.StdEncoding.EncodeToString(state.Bytes())
+		importPayload := fmt.Sprintf("%s/%s", stateID, modelEncoded)
 
 		for _, providerConfig := range providerConfigs {
+			// given
+			config := providerConfig +
+				builders.NewResourceTFBuilder().
+					WithName("test1").
+					WithMethod("GET").
+					WithPath("/posts/1").
+					Build()
+
+			// when
 			resource.UnitTest(t, resource.TestCase{
 				PreCheck:                 func() { testAccPreCheck(t) },
 				ProtoV6ProviderFactories: testAccProtoV6ProviderFactories,
 				Steps: []resource.TestStep{
-					// create and read testing for GET method
+					// Apply testing
 					{
-						Config: providerConfig +
-							builders.NewResourceTFBuilder().
-								WithName("test1").
-								WithMethod("GET").
-								WithPath("/posts/1").
-								Build(),
+						Config: config,
 						Check: resource.ComposeAggregateTestCheckFunc(
+							// then
 							resource.TestCheckResourceAttr("http_request.test1", "method", "GET"),
 							resource.TestCheckResourceAttr("http_request.test1", "path", "/posts/1"),
-							// Verify dynamic values have any value set in the state.
 							resource.TestCheckResourceAttrSet("http_request.test1", "id"),
 							resource.TestCheckResourceAttrSet("http_request.test1", "response_code"),
 							resource.TestCheckResourceAttrSet("http_request.test1", "response_body"),
 						),
 					},
 
-					// ImportState testing for GET method
+					// Destroy testing
 					{
-						ResourceName:            "http_request.test1",
-						ImportState:             true,
-						ImportStateVerify:       true,
-						ImportStateId:           state1ID,
-						ImportStateVerifyIgnore: []string{"is_response_body_json"}, // bool empty are considered null via Terraform SDK
+						Destroy: true,
+						Config:  config,
+					},
+
+					// Import testing
+					{
+						ImportState:   true,
+						ResourceName:  "http_request.test1",
+						ImportStateId: importPayload,
+						// function is being used because ImportStateVerify compares with the previous object
+						ImportStateCheck: func(state []*terraform.InstanceState) error {
+							// then
+							assert.Equal(t, stateID, state[0].ID, "id should be equal to the stateID")
+							assert.Equal(t, "GET", state[0].Attributes["method"], "method should be GET")
+							assert.Equal(t, "/posts/1", state[0].Attributes["path"], "path should be /posts/1")
+							assert.Equal(t, "200", state[0].Attributes["response_code"], "response_code should be 200")
+							return nil
+						},
 					},
 				},
 			})
@@ -86,59 +108,78 @@ func TestHTTPRequestResource(t *testing.T) {
 	})
 
 	t.Run("should apply and check the state when using POST and a non-JSON request body", func(t *testing.T) {
-		var state2 bytes.Buffer
-		_ = json.Compact(&state2, []byte(`{
+		// given
+		var state bytes.Buffer
+		_ = json.Compact(&state, []byte(`{
 		"method": "POST",
 		"path": "/posts",
 		"request_body": "test body",
 		"response_code": 201,
 		"response_body":"{\n  \"id\": 101\n}"
 	}`))
-		state2ID := base64.StdEncoding.EncodeToString(state2.Bytes())
+		stateID := "anything unique"
+		modelEncoded := base64.StdEncoding.EncodeToString(state.Bytes())
+		importPayload := fmt.Sprintf("%s/%s", stateID, modelEncoded)
 
 		for _, providerConfig := range providerConfigs {
+			// given
+			config := providerConfig +
+				builders.NewResourceTFBuilder().
+					WithName("test2").
+					WithMethod("POST").
+					WithPath("/posts").
+					WithRequestBody("\"test body\"").
+					Build()
+
+			// when
 			resource.UnitTest(t, resource.TestCase{
 				PreCheck:                 func() { testAccPreCheck(t) },
 				ProtoV6ProviderFactories: testAccProtoV6ProviderFactories,
 				Steps: []resource.TestStep{
-					// create and read testing for POST method
+					// Apply testing
 					{
-						Config: providerConfig +
-							builders.NewResourceTFBuilder().
-								WithName("test2").
-								WithMethod("POST").
-								WithPath("/posts").
-								WithRequestBody("\"test body\"").
-								Build(),
+						Config: config,
 						Check: resource.ComposeAggregateTestCheckFunc(
+							// then
 							resource.TestCheckResourceAttr("http_request.test2", "method", "POST"),
 							resource.TestCheckResourceAttr("http_request.test2", "path", "/posts"),
 							resource.TestCheckResourceAttr("http_request.test2", "request_body", "test body"),
-							// Verify dynamic values have any value set in the state.
-							resource.TestCheckResourceAttr("http_request.test2", "id", state2ID),
+							resource.TestCheckResourceAttrSet("http_request.test2", "id"),
 							resource.TestCheckResourceAttr("http_request.test2", "response_code", "201"),
 							resource.TestCheckResourceAttrSet("http_request.test2", "response_body"),
 						),
 					},
 
-					// ImportState testing for POST method
+					// Destroy testing
 					{
-						ResourceName:            "http_request.test2",
-						ImportState:             true,
-						ImportStateVerify:       true,
-						ImportStateId:           state2ID,
-						ImportStateVerifyIgnore: []string{"is_response_body_json"}, // bool empty are considered null via Terraform SDK
+						Destroy: true,
+						Config:  config,
 					},
 
-					// TODO: update and read testing should be implemented here
+					// Import testing
+					{
+						ImportState:   true,
+						ResourceName:  "http_request.test2",
+						ImportStateId: importPayload,
+						// function is being used because ImportStateVerify compares with the previous object
+						ImportStateCheck: func(state []*terraform.InstanceState) error {
+							// then
+							assert.Equal(t, stateID, state[0].ID, "id should be equal to the stateID")
+							assert.Equal(t, "POST", state[0].Attributes["method"], "method should be POST")
+							assert.Equal(t, "/posts", state[0].Attributes["path"], "path should be /posts")
+							assert.Equal(t, "201", state[0].Attributes["response_code"], "response_code should be 201")
+							return nil
+						},
+					},
 				},
 			})
 		}
 	})
 
 	t.Run("should apply and check the state when using POST and a JSON request body", func(t *testing.T) {
-		var state3 bytes.Buffer
-		_ = json.Compact(&state3, []byte(`{
+		// given
+		var state bytes.Buffer
+		_ = json.Compact(&state, []byte(`{
 		"method": "POST",
 		"path": "/posts",
 		"request_body": "{ \"test\": \"test body\" }",
@@ -149,34 +190,40 @@ func TestHTTPRequestResource(t *testing.T) {
 		"response_body_id": "101",
 		"response_body_json": {"id":"101"}
 	}`))
-		state3ID := base64.StdEncoding.EncodeToString(state3.Bytes())
+		stateID := "anything unique"
+		modelEncoded := base64.StdEncoding.EncodeToString(state.Bytes())
+		importPayload := fmt.Sprintf("%s/%s", stateID, modelEncoded)
 
 		body, _ := json.Marshal("{ \"test\": \"test body\" }")
 
 		for _, providerConfig := range providerConfigs {
+			// given
+			config := providerConfig +
+				builders.NewResourceTFBuilder().
+					WithName("test3").
+					WithMethod("POST").
+					WithPath("/posts").
+					WithRequestBody(string(body)).
+					WithResponseBodyIDFilter("$.id").
+					WithIsResponseBodyJSON(true).
+					Build()
+
+			// when
 			resource.UnitTest(t, resource.TestCase{
 				PreCheck:                 func() { testAccPreCheck(t) },
 				ProtoV6ProviderFactories: testAccProtoV6ProviderFactories,
 				Steps: []resource.TestStep{
-					// create and read testing for POST method with response filtering
+					// Apply testing
 					{
-						Config: providerConfig +
-							builders.NewResourceTFBuilder().
-								WithName("test3").
-								WithMethod("POST").
-								WithPath("/posts").
-								WithRequestBody(string(body)).
-								WithResponseBodyIDFilter("$.id").
-								WithIsResponseBodyJSON(true).
-								Build(),
+						Config: config,
 						Check: resource.ComposeAggregateTestCheckFunc(
+							// then
 							resource.TestCheckResourceAttr("http_request.test3", "method", "POST"),
 							resource.TestCheckResourceAttr("http_request.test3", "path", "/posts"),
 							resource.TestCheckResourceAttr("http_request.test3", "request_body", "{ \"test\": \"test body\" }"),
 							resource.TestCheckResourceAttr("http_request.test3", "is_response_body_json", "true"),
 							resource.TestCheckResourceAttr("http_request.test3", "response_body_id_filter", "$.id"),
-							// Verify dynamic values have any value set in the state.
-							resource.TestCheckResourceAttr("http_request.test3", "id", state3ID),
+							resource.TestCheckResourceAttrSet("http_request.test3", "id"),
 							resource.TestCheckResourceAttr("http_request.test3", "response_code", "201"),
 							resource.TestCheckResourceAttrSet("http_request.test3", "response_body"),
 							resource.TestCheckResourceAttr("http_request.test3", "response_body_id", "101"),
@@ -184,15 +231,30 @@ func TestHTTPRequestResource(t *testing.T) {
 						),
 					},
 
-					// ImportState testing for POST method with response filtering
+					// Destroy testing
 					{
-						ResourceName:            "http_request.test3",
-						ImportState:             true,
-						ImportStateVerify:       true,
-						ImportStateVerifyIgnore: []string{"is_response_body_json"}, // bool empty are considered null via Terraform SDK
+						Destroy: true,
+						Config:  config,
 					},
 
-					// TODO: update and read testing should be implemented here
+					// Import testing
+					{
+						ImportState:   true,
+						ResourceName:  "http_request.test3",
+						ImportStateId: importPayload,
+						// function is being used because ImportStateVerify compares with the previous object
+						ImportStateCheck: func(state []*terraform.InstanceState) error {
+							// then
+							assert.Equal(t, stateID, state[0].ID, "id should be equal to the stateID")
+							assert.Equal(t, "POST", state[0].Attributes["method"], "method should be POST")
+							assert.Equal(t, "/posts", state[0].Attributes["path"], "path should be /posts")
+							assert.Equal(t, "201", state[0].Attributes["response_code"], "response_code should be 201")
+							assert.Equal(t, "{ \"test\": \"test body\" }", state[0].Attributes["request_body"], "request_body should be { \"test\": \"test body\" }")
+							assert.Equal(t, "true", state[0].Attributes["is_response_body_json"], "is_response_body_json should be true")
+							assert.Equal(t, "$.id", state[0].Attributes["response_body_id_filter"], "response_body_id_filter should be $.id")
+							return nil
+						},
+					},
 				},
 			})
 		}